KafkaDocker/consumer/requirements.txt
kafka-python
jason
